cambios en el status de las tareas

+++ usuarios/views.py
@@ -84,6 +84,17 @@ class AgregarUsuario(View):
             #     messages.error(request, form.error_messages[msg])
             
             return render(request,self.template_name,{"form":form, "form2":form2})
+        
+    @method_decorator(login_required)             #Puede agrupar decoradores
+    def dispatch(self, request, *args, **kwargs):
+        
+        if not request.user.has_perm('auth.add_user'):
+              messages.success(request,"ACCESO DENEGADO")
+              return redirect('lista_usuarios')
+
+        return super(AgregarUsuario,self).dispatch(request, *args, **kwargs)
+
+
     
 
 class EditarUsuario(UpdateView):
@@ -101,8 +112,15 @@ class EditarUsuario(UpdateView):
 
 
 
+@login_required
 def EliminarUsuario(request, pk):
-    user = get_object_or_404(User, id = pk)
-    user.delete()
-    messages.success(request,"Usuario eliminado")
-    return redirect('lista_usuarios')
+
+
+    if not request.user.has_perm('auth.delete_user'):
+              messages.success(request,"ACCESO DENEGADO")
+              return redirect('lista_usuarios')
+    else:
+        user = get_object_or_404(User, id = pk)
+        user.delete()
+        messages.success(request,"Usuario eliminado")
+        return redirect('lista_usuarios')
